Changed to 16bit colors <3

+++ Yyunko/Chritmiss18/GlowyFaceCode/components/NeoController/include/Color.h
@@ -37,10 +37,10 @@ public:
 		uint8_t b;
 	};
 
-	uint8_t r;
-	uint8_t g;
-	uint8_t b;
-	uint8_t alpha;
+	uint16_t r;
+	uint16_t g;
+	uint16_t b;
+	uint16_t alpha;
 
 	Color();
 	Color(uint32_t cCode, uint8_t brightness = 255);
